Tests, spec, refactoring

+++ app/controllers/boards_controller.rb
@@ -1,12 +1,17 @@
+
 class BoardsController < ApplicationController
-#before_action :authenticate_user!
+before_action :authenticate_user!
 
 
 def index
 
-    #check_auth
+    check_auth
 
-    boards = Board.all
+    #boards = Board.all
+    
+
+    # Only render the boards for the current user
+    boards = Board.where(user_id: current_user.id)
     render json: boards, status: :ok
 
   end
@@ -14,40 +19,66 @@ def index
   
   def show
 
-    #check_auth
+    check_auth
 
-    board = Board.find_by(id: params[:id])
+    if !params[:id].nil?
+    
+      board = Board.find_by(id: params[:id], user_id: current_user.id)
+    end
+
+
     if board
-      render json: board, status: :ok
+      
+        render json: board, status: :ok
+      
     else
       render json: {
-        error: "Board not found"
+        error: "Board not found!"
       }
     end
   end
 
   def create
 
-    #check_auth
+    check_auth
 
-    board = Board.new(board_params)
+    board = current_user.boards.build(board_params)
+    #board = Board.new(board_params)
     #board.user_id = current_user.id
+
+    puts "CREATE"
+    puts board.inspect
     
     if board.save
       render json: board, status: :created
     else
-      render json: board.errors, status: :unprocessable_entity
+      render json: {errors: board.errors}, status: :unprocessable_entity
     end
   end
 
   def update
 
-    #check_auth
+    check_auth
 
     board = Board.find_by(id: params[:id])
     if board
-      board.update(title: params[:title])
-      render json: "Board Record Updated Successfully"
+      # Check
+      if board.user_id == current_user.id
+
+        success = board.update(title: params[:title])
+
+        if success
+          render json: "Board Record Updated Successfully"
+        elsif board.errors.present?
+          render json: {errors: board.errors}, status: :unprocessable_entity
+        end
+
+      else
+        render json: {
+        error: "Board not yours!"
+      }
+      end
+
     else
       render json: {
         error: "Board Not Found"
@@ -57,12 +88,19 @@ def index
 
   def destroy
 
-    #check_auth
+    check_auth
 
     board = Board.find_by(id: params[:id])
     if board
-      board.destroy
-      render json: "Board Has Been Deleted"
+      # Check
+      if board.user_id == current_user.id
+        board.destroy
+        render json: "Board Has Been Deleted"
+      else
+        render json: {
+        error: "Board not yours!"
+      }
+      end 
     else
       render json: {
         error: "Board not Found"
@@ -78,12 +116,12 @@ def index
 
 
 
-  # def check_auth
-  #   if user_signed_in? == false
-  #     render json: {
-  #       message: "User not authenticated"
-  #     }
-  #   end
-  # end
+  def check_auth
+    if !user_signed_in?
+      render json: {
+        message: "User not authenticated"
+      }
+    end
+  end
 end
